LayoutTests/imported/w3c/web-platform-tests/fullscreen/api/document-exit-fullscreen-nested-shadow-dom-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: outer.requestFullscreen is not a function. (In 'outer.requestFullscreen()', 'outer.requestFullscreen' is undefined)
fullscreen

Harness Error (FAIL), message = Unhandled rejection: outer.requestFullscreen is not a function. (In 'outer.requestFullscreen()', 'outer.requestFullscreen' is undefined)

TIMEOUT Exiting fullscreen from a nested shadow root works correctly. Test timed out

Harness Error (FAIL), message = Unhandled rejection: outer.requestFullscreen is not a function. (In 'outer.requestFullscreen()', 'outer.requestFullscreen' is undefined)

TIMEOUT Exiting fullscreen from a nested shadow root works correctly. Test timed out
